new room cleaned endpoint
the regular put save method wasn't going to cut it so I quick whipped up a specialized way to do it and tested it to work. Plus I tested the add bulk rooms function and that works as well. From where I'm sitting we are all ready to record and call it done.