tinker with delete error
 No matter what I do it keeps screaming at me that it can't delete a parent with a foreign key associated with it even though I have everything set to CascadeType.PERIST and MERGE!!!

Within my delete method in the service layer I even manually disconnect the housekeeper from any of the other entities it might be connected to and them from it. Then delete it but it's still giving me the same thing. It's making me think I don't understand Many to many relationship annotations.

cause I tested to make sure it was actually separating all the entities when I asked and it is. But something hidden somewhere is making my dao uncomfortable.